Store location and rotation in DisplayObject
The DisplayObject struct now contains glm::vec3s for the object's
location and rotation. The createMVP method is now updateMVP, and takes
its parameters from a passed object, attaching the resulting matrix set
to the object.

To test this, the clanger model is loaded and displayed next to the
crate.